sequelized all the backend

+++ backend/controllers/albumController.js
@@ -1,31 +1,52 @@
 const fs = require('fs')
 const Album = require('../models/Album')
+const AlbumP = require('../models/AlbumP')
 const Entity = require('../models/Entity')
+const EntityP = require('../models/EntityP')
+const FileP = require('../models/FileP')
 const { deleteFile, deleteFiles } = require('../utils/deleteFile')
 
 const { BadRequest, NotFound, Unauthorized } = require('../utils/errors')
 const { albumValidator } = require('../validators/albumValidator')
 
+const albumIncludes = [
+  {
+    model: FileP,
+    attributes: ['filename', 'filepath'],
+  },
+  {
+    model: EntityP,
+    attributes: ['id', 'name', 'alias'],
+  },
+]
+
 module.exports.postAlbum = async (req, res, next) => {
   const { isAdmin, isManager, isModerator, roles } = req.user
   const { id: albumId, action, entityAlias } = req.query
+  const filename = req.file ? req.file.filename : null
+  const filepath = req.file ? req.file.path : null
 
-  const entity = await Entity.findOne({ alias: entityAlias })
-  if (!entity) {
-    if (action !== 'delete')
+  if (action !== 'delete' && !entityAlias)
+    return next(new BadRequest('entityAlias missing'))
+  const entity =
+    action !== 'delete'
+      ? await EntityP.findOne({ where: { alias: entityAlias } })
+      : null
+
+  if (action !== 'delete') {
+    if (!entity)
       return next(
         deleteFile(
-          req.file.path,
+          filepath,
           new BadRequest('une ou plusieurs données manquantes: entité')
         )
       )
-    return next(new BadRequest('une ou plusieurs données manquantes: entité'))
   }
 
   let roleIsAllowed = false
 
   roles.map((role) => {
-    if (JSON.stringify(entity._id === role.entity)) {
+    if (JSON.stringify(entity.id === role.entity)) {
       roleIsAllowed = true
     }
   })
@@ -54,8 +75,6 @@ module.exports.postAlbum = async (req, res, next) => {
     if (!req.file)
       return next(new BadRequest('Please select an image to upload'))
 
-    const { filename, path: filepath } = req.file
-
     if (!filepath || !filename)
       return next(
         new BadRequest(
@@ -72,15 +91,17 @@ module.exports.postAlbum = async (req, res, next) => {
         )
       )
 
-    const album = req.body
-
-    const newAlbum = new Album(album)
-    newAlbum.coverpath = filepath
-    newAlbum.covername = filename
-    newAlbum.entity = entity._id
+    const newAlbum = AlbumP.build(req.body)
+    newAlbum.entityId = entity.id
     try {
       const savedAlbum = await newAlbum.save()
       if (savedAlbum) {
+        await FileP.create({
+          filename,
+          filepath,
+          type: 'image',
+          albumId: savedAlbum.id,
+        })
         return res.status(201).send({ message: 'Album correctement créee' })
       }
     } catch (err) {
@@ -89,28 +110,25 @@ module.exports.postAlbum = async (req, res, next) => {
   } else if (action === 'update' && albumId) {
     // case update
 
-    const currentAlbum = await Album.findOne({ _id: albumId })
+    const currentAlbum = await AlbumP.findOne({
+      where: { id: albumId },
+      include: albumIncludes,
+    })
     if (!currentAlbum) return next(new BadRequest("L'album n'existe pas"))
-    if (req.file) {
-      fs.unlink(currentAlbum.coverpath, (err) => {
-        if (err) return next(err)
-      })
+
+    const updates = {
+      name: req.body.name || currentAlbum.name,
+      description: req.body.description || currentAlbum.description,
+      isPrivate: req.body.isPrivate,
+      covername: req.file ? req.file.filename : currentAlbum.covername,
+      coverpath: req.file ? req.file.path : currentAlbum.coverpath,
     }
 
     try {
-      const updatedAlbum = await Album.findOneAndUpdate(
-        { _id: albumId },
-        {
-          name: req.body.name || currentAlbum.name,
-          description: req.body.description || currentAlbum.description,
-          isPrivate: req.body.isPrivate,
-          covername: req.file ? req.file.filename : currentAlbum.covername,
-          coverpath: req.file ? req.file.path : currentAlbum.coverpath,
-        },
-        {
-          returnOriginal: false,
-        }
-      )
+      const updatedAlbum = await AlbumP.update(updates, {
+        where: { id: albumId },
+      })
+
       if (updatedAlbum) {
         if (process.env.NODE_ENV === 'production') {
           return res.status(200).send('Album correctement modifiée')
@@ -121,34 +139,30 @@ module.exports.postAlbum = async (req, res, next) => {
         })
       }
     } catch (err) {
-      return next(deleteFile(req.file.path, err))
+      return next(deleteFile(null, err))
     }
   } else if (action === 'delete' && albumId) {
     try {
-      const deletedChemin = await Album.findOneAndDelete({ _id: albumId })
+      const toDeleteAlbum = await AlbumP.findOne({
+        where: { id: albumId },
+        include: albumIncludes,
+      })
+      if (!toDeleteAlbum)
+        return res.status(200).send('album avait deja été supprimé')
 
-      // delete coverpath
-      if (deletedChemin) {
-        fs.unlink(deletedChemin.coverpath, (err) => {
-          if (err) return next(err)
+      const deletionImageErrors = []
+      toDeleteAlbum.files.forEach((file) => {
+        fs.unlink(file.filepath, (err) => {
+          if (err) deletionImageErrors.push(err)
         })
+      })
 
-        // delete all images pah in the album
+      if (deletionImageErrors.length > 0)
+        return next(deletionImageErrors.join())
 
-        if (deletedChemin.pictures.length > 0) {
-          const deletionErrors = []
-          for (let i = 0; i < deletedChemin.pictures.length; i += 1) {
-            const picture = deletedChemin.pictures[i]
-            fs.unlink(picture.filepath, (err) => {
-              if (err) deletionErrors.push(err)
-            })
-          }
-
-          if (deletionErrors.length > 0) return next(deletionErrors.join())
-        }
-
+      const deletedAlbum = await toDeleteAlbum.destroy()
+      if (deletedAlbum)
         return res.status(200).send({ message: 'album correctement effacé' })
-      }
     } catch (err) {
       return next(err)
     }
@@ -163,23 +177,22 @@ module.exports.getAlbums = async (req, res, next) => {
     return next(new BadRequest(errors.join()))
   }
 
-  if (req.query.id) {
-    req.query._id = req.query.id
-    delete req.query.id
-  }
   if (req.query.entityAlias) {
-    const entity = await Entity.findOne({ alias: req.query.entityAlias })
+    const entity = await EntityP.findOne({
+      where: { alias: req.query.entityAlias },
+    })
     if (!entity) return next(new BadRequest("cette entité n'existe pas"))
-    req.query.entity = entity._id
+    req.query.entityId = entity.id
     delete req.query.entityAlias
   }
 
   try {
-    const albums = await Album.find(req.query)
-      .select(
-        '_id name alias description entity coverpath covername pictures isPrivate'
-      )
-      .populate('entity')
+    const albums = await AlbumP.findAll({
+      where: req.query,
+      attributes: ['id', 'name', 'alias'],
+      limit: 10,
+      include: albumIncludes,
+    })
 
     if (albums.length < 1)
       return next(
@@ -195,7 +208,13 @@ module.exports.postAlbumImages = async (req, res, next) => {
   const { isAdmin, isManager, isModerator, roles } = req.user
   const { id: albumId, action, entityAlias, filepath } = req.query
 
-  const entity = await Entity.findOne({ alias: entityAlias })
+  if (action !== 'delete' && (!entityAlias || !albumId))
+    return next(new BadRequest('params missing: entityAlias, albumId, action '))
+
+  const entity =
+    action === 'delete'
+      ? null
+      : await EntityP.findOne({ where: { alias: entityAlias } })
   // if (!entity)
   //   return next(
   //     deleteFiles(
@@ -207,7 +226,7 @@ module.exports.postAlbumImages = async (req, res, next) => {
   let roleIsAllowed = false
 
   roles.map((role) => {
-    if (JSON.stringify(entity._id === role.entity)) {
+    if (JSON.stringify(entity.id === role.entity.id)) {
       roleIsAllowed = true
     }
   })
@@ -229,41 +248,42 @@ module.exports.postAlbumImages = async (req, res, next) => {
     return next(new BadRequest('les images ne sont pas telechargées'))
 
   // check if album exist
-  const album = await Album.findOne({ _id: albumId })
+  const album = await AlbumP.findOne({
+    where: { id: albumId },
+    includes: albumIncludes,
+  })
 
   if (!album)
     return next(deleteFiles(req.files, new BadRequest('album non identifié')))
 
-  const { pictures } = album
-
   if (action === 'create') {
-    for (let i = 0; i < req.files.length; i += 1) {
-      const { filename, path } = req.files[i]
-      pictures.push({ filename, filepath: path })
-    }
-
     try {
-      const modifiedAlbum = await album.save()
-      if (modifiedAlbum)
-        return res.status(200).send({ message: 'images enregistrées' })
+      req.files.forEach(async (file) => {
+        const { filename, path } = file
+        await FileP.create({
+          filename,
+          filepath: path,
+          type: 'image',
+          albumId: album.id,
+        })
+      })
+
+      return res.status(200).send({ message: 'images enregistrées' })
     } catch (err) {
       return next(deleteFiles(req.files, new BadRequest(err)))
     }
   } else if (action === 'delete') {
-    const filteredPictures = pictures.filter(
-      ({ filepath: path }) => path !== filepath
-    )
-
     try {
-      fs.unlink(filepath, (err) => {
-        if (err) return err
+      const deletedFile = await FileP.destroy({
+        where: { albumId: albumId, filepath },
       })
-      const modifiedAlbum = await Album.findOneAndUpdate(
-        { _id: albumId },
-        { pictures: filteredPictures }
-      )
-      if (modifiedAlbum)
+
+      if (deletedFile) {
+        fs.unlink(filepath, async (err) => {
+          if (err) return err
+        })
         return res.status(200).send({ message: 'image supprimée' })
+      }
     } catch (err) {
       return next(new BadRequest(err))
     }
